@@ -1,36 +1,36 @@
 {
   "name": "ng-workshop",
-  "version": "5.0.0-rc2",
+  "version": "5.0.1",
   "private": true,
   "scripts": {
     "start": "\"ng build workshop --prod --watch\""
   },
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
+    "@angular/animations": "^8.0.2",
+    "@angular/common": "^8.0.2",
+    "@angular/compiler": "^8.0.2",
+    "@angular/core": "^8.0.2",
+    "@angular/forms": "^8.0.2",
     "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/platform-browser": "^8.0.2",
+    "@angular/platform-browser-dynamic": "^8.0.2",
+    "@angular/router": "^8.0.2",
     "@ngx-translate/core": "11.0.1",
     "@ngx-translate/http-loader": "4.0.0",
-    "core-js": "^2.5.7",
+    "core-js": "^3.1.4",
     "rxjs": "^6.5.2",
     "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/build-ng-packagr": "~0.13.9",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@types/jasmine": "~3.3.12",
+    "@angular-devkit/build-angular": "^0.800.4",
+    "@angular-devkit/build-ng-packagr": "~0.800.4",
+    "@angular/cli": "~8.0.4",
+    "@angular/compiler-cli": "^8.0.2",
+    "@angular/language-service": "^8.0.2",
+    "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.0",
-    "codelyzer": "~5.0.1",
+    "@types/node": "~12.0.10",
+    "codelyzer": "~5.1.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.1.0",
@@ -38,12 +38,12 @@
     "karma-coverage-istanbul-reporter": "~2.0.5",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
-    "ng-packagr": "^5.1.0",
+    "ng-packagr": "^5.3.0",
     "protractor": "^5.4.2",
-    "ts-node": "~8.1.0",
+    "ts-node": "~8.3.0",
     "tsickle": ">=0.35.0",
-    "tslib": "^1.9.3",
-    "tslint": "~5.16.0",
-    "typescript": "3.1.6"
+    "tslib": "^1.10.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.4.0"
   }
 }
