@@ -1,6 +1,6 @@
 {
   "name": "@softheon/ng-workshop",
-  "version": "5.0.0-rc2",
+  "version": "5.0.1",
   "description": "A suite of modern Angular components built on the Softheon Workshop Design System",
   "repository": {
     "type": "git",
@@ -12,7 +12,7 @@
     "url": "https://hack.softheon.io/Workshop/"
   },
   "peerDependencies": {
-    "@angular/common": "^7.2.15",
-    "@angular/core": "^7.2.15"
+    "@angular/common": "^8.0.2",
+    "@angular/core": "^8.0.2"
   }
 }
